Refactor running of node to evaluate, fuzz, and detect
