+++ graphqler/fuzzer/engine/detectors/__init__.py
@@ -1,4 +1,4 @@
-from .os_command_injection_detector import OSCommandInjectionDetector
+from .os_command_injection.os_command_injection_detector import OSCommandInjectionDetector
 
 injection_detectors = [
     OSCommandInjectionDetector
