+++ graphqler/fuzzer/engine/detectors/path_injection/path_traversal_injection_materializer.py
@@ -1,24 +1,26 @@
-from .injection_materializer import InjectionMaterializer
-from graphqler.utils.api import API
-from ..getter import Getter
-
 from typing import override
+
+from graphqler.utils.api import API
+
+from ...materializers.getter import Getter
+from ...materializers.injection_materializer import InjectionMaterializer
 
 
 # The main class that's being used
-class SQLInjectionMaterializer(InjectionMaterializer):
+class PathTraversalInjectionMaterializer(InjectionMaterializer):
     def __init__(self, api: API, fail_on_hard_dependency_not_met: bool = False, max_depth: int = 20):
         super().__init__(api, fail_on_hard_dependency_not_met)
         self.api = api
         self.fail_on_hard_dependency_not_met = fail_on_hard_dependency_not_met
-        self.getter = SQLInjectionGetter()
+        self.max_depth = max_depth
+        self.getter = PathTraversalInjectionGetter()
 
 
 # Override the getters class to add custom getters for SQL injection
-class SQLInjectionGetter(Getter):
+class PathTraversalInjectionGetter(Getter):
     def __init__(self):
         super().__init__()
 
     @override
     def get_random_string(self, input_name: str) -> str:
-        return '"aaa \' OR 1=1--"'
+        return '"../../../../etc/passwd"'
