+++ graphqler/fuzzer/engine/materializers/__init__.py
@@ -6,11 +6,9 @@ from .regular_payload_materializer import RegularPayloadMaterializer
 
 # Attack materializers
 from .dos import dos_materializers
-from .injection import injection_materializers
 
 __all__ = [
     "Materializer",
     "RegularPayloadMaterializer",
     "dos_materializers",
-    "injection_materializers",
 ]
